requirements.txt
langgraph
langchain_community
langchain_openai
openai
chromadb
sentence-transformers
langchain-community
fastapi[standard]